Nuam_Backend/core/urls.py
from django.contrib import admin
from django.urls import path
from .views import MyTokenObtainPairView
from rest_framework_simplejwt.views import (TokenRefreshView, TokenVerifyView)

urlpatterns = [
    path('token/', MyTokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
]
