Nuam_Backend/requirements.txt
Django>=5.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
djoser
psycopg2-binary>=2.9.9

